added logic to retrieve the participant info from the database\n based on their email, and if they dont yet exist, add them to the database as a new participant

+++ db/schema.sql
@@ -5,7 +5,7 @@ CREATE TABLE participants (
   subject smallserial NOT NULL,
   email character varying(40) UNIQUE NOT NULL,
   sessions_completed smallint NOT NULL,
-  rng_seed integer NOT NULL,
+  rng_seed bigint NOT NULL,
   PRIMARY KEY(subject)
 );
